Move functions that don't return a url outside of apiUrls.
